requirements.txt
# Runtime dependencies — pinned to LTS / current stable
Django>=5.2,<5.3
djangorestframework>=3.16,<3.17
django-environ>=0.12,<0.13
django-htmx>=1.21,<2.0
psycopg[binary]>=3.2,<3.3
redis>=5.2,<6.0
gunicorn>=23.0,<24.0
whitenoise>=6.8,<7.0
celery[redis]>=5.4,<5.5
python-telegram-bot>=21.0,<22.0
